+++ backend/src/main/java/org/cathori/backend/alert/application/DispatchNotificationService.java
@@ -3,9 +3,7 @@ package org.cathori.backend.alert.application;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 import org.cathori.backend.alert.domain.AlertHistory;
 import org.cathori.backend.alert.domain.AlertHistoryRepository;
@@ -22,7 +20,7 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class AlertService {
+public class DispatchNotificationService {
 
     private final AlertHistoryRepository alertHistoryRepository;
     private final AlertResultWriter alertResultWriter;
@@ -40,20 +38,6 @@ public class AlertService {
         for (Notice notice : unsentNotices) {
             log.info("공지id {} 제목 {} 발송 시작", notice.getId(), notice.getTitle());
             dispatchForNotice(notice);
-        }
-    }
-
-    public void retryFailedAlerts() {
-        List<AlertHistory> failedLogs = alertHistoryRepository.findFailedForRetry();
-        if (failedLogs.isEmpty()) return;
-
-        log.info("FCM 재시도 {}건", failedLogs.size());
-
-        Map<Long, List<AlertHistory>> byNotice = failedLogs.stream()
-                .collect(Collectors.groupingBy(AlertHistory::getNoticeId));
-
-        for (Map.Entry<Long, List<AlertHistory>> entry : byNotice.entrySet()) {
-            retryForNotice(entry.getKey(), entry.getValue());
         }
     }
 
@@ -103,36 +87,6 @@ public class AlertService {
         alertResultWriter.persistDispatchResult(notice, successIds, failedIds);
         
         log.info("FCM 발송 완료. noticeId={}, success={}, fail={}", notice.getId(), successIds.size(), failedIds.size());
-    }
-
-    private void retryForNotice(Long noticeId, List<AlertHistory> logs) {
-        Notice notice = noticeRepository.findById(noticeId).orElse(null);
-        if (notice == null) return;
-
-        List<UserToken> targets = logs.stream()
-                .map(l -> {
-                    User user = userRepository.findById(l.getUserId()).orElse(null);
-                    if (user == null || user.getDeviceToken() == null) return null;
-                    return new UserToken(l.getUserId(), user.getDeviceToken());
-                })
-                .filter(Objects::nonNull)
-                .toList();
-
-        if (targets.isEmpty()) return;
-
-        List<UserFcmResult> results;
-        try {
-            results = fcmPort.sendMulticast(targets, "Cathori 새 공지", notice.getTitle(), noticeId);
-        } catch (Exception e) {
-            log.error("FCM 재시도 실패. noticeId={}", noticeId, e);
-            List<Long> userIds = targets.stream().map(UserToken::userId).toList();
-            alertResultWriter.persistRetryFailure(noticeId, userIds);
-            return;
-        }
-
-        List<Long> successIds = results.stream().filter(UserFcmResult::success).map(UserFcmResult::userId).toList();
-        List<Long> failedIds = results.stream().filter(r -> !r.success()).map(UserFcmResult::userId).toList();
-        alertResultWriter.persistRetryResult(noticeId, successIds, failedIds);
     }
 
     private List<UserToken> buildTargets(Notice notice) {
