+++ backend/src/main/java/org/cathori/backend/alert/infra/AlertRetryScheduler.java
@@ -1,7 +1,7 @@
 package org.cathori.backend.alert.infra;
 
 import lombok.RequiredArgsConstructor;
-import org.cathori.backend.alert.application.AlertService;
+import org.cathori.backend.alert.application.RetryDispatchNotificationService;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
@@ -15,11 +15,11 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class AlertRetryScheduler {
 
-    private final AlertService alertService;
+    private final RetryDispatchNotificationService retryDispatchNotificationService;
 
     /** 20분마다 실패한 알림 발송을 재시도합니다. */
     @Scheduled(cron = "0 0/20 * * * *")
     public void retryFailedAlerts() {
-        alertService.retryFailedAlerts();
+        retryDispatchNotificationService.retryFailedAlerts();
     }
 }
